% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table.R
\name{clean_table}
\alias{clean_table}
\title{Clean the table and add the text of the speechs}
\usage{
clean_table(tab, txt)
}
\arguments{
\item{tab}{durty table}

\item{txt}{vector with the texts}
}
\value{
return the cleaned table
}
\description{
Clean the columns names, exclude empty rows and separate columns.
}
\keyword{internal}
