% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table.R
\name{extract_table}
\alias{extract_table}
\title{Get the table with informations}
\usage{
extract_table(r_html)
}
\arguments{
\item{r_html}{a response HTML file}
}
\value{
return the table with informations about the speaker
}
\description{
Get the table as a tibble with the information
}
\keyword{internal}
