% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-speech-text.R
\name{get_speech_text}
\alias{get_speech_text}
\title{Obtains the text of the speech}
\usage{
get_speech_text(txt_urls)
}
\arguments{
\item{txt_urls}{URL's to read HTML}
}
\value{
vector of text
}
\description{
Extract the text of the speech given an URL
}
\keyword{internal}
