% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table.R
\name{num_pag}
\alias{num_pag}
\title{Calculate the number of pages that the request has.}
\usage{
num_pag(r_html)
}
\arguments{
\item{r_html}{a response HTML file}
}
\value{
return the number of pages that exist
}
\description{
Extract the number of pages and calculate the specif number of pages for avoid future errors.
}
\keyword{internal}
