% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-speech-table.R
\name{speech_data}
\alias{speech_data}
\title{Obtains the text of the speech}
\usage{
speech_data(keyword)
}
\arguments{
\item{keyword}{argument of the \code{download_page} function.}
}
\value{
vector the speech data with all informational columns and the speech.
}
\description{
Extract the text of the speech given an URL.
}
\examples{
speech_data(keyword = "pandemia")
speech_data(keyword = '"meio ambiente"')

}
