% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-speech-text.R
\name{transformer_url}
\alias{transformer_url}
\title{Obtains the text of the speech}
\usage{
transformer_url(r_html)
}
\arguments{
\item{r_html}{a response HTML file}
}
\value{
a transformed URL
}
\description{
Extract the text of the speech given an URL
}
\keyword{internal}
