% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-table.R
\name{download_page}
\alias{download_page}
\title{Obtains the HTML file of a current page given a main text or phrase.}
\usage{
download_page(
  tx_text,
  current_page = 1,
  reference_date = "2021-12-20",
  qtd_days = 3
)
}
\arguments{
\item{tx_text}{principal text or phrase present on speech}

\item{current_page}{current page of search (integer numbers)}

\item{reference_date}{end date of search}

\item{qtd_days}{quantity of days before the \code{reference_date}}
}
\value{
invisible, return a HTML file
}
\description{
GET request of the pages of \verb{DISCURSOS E NOTAS TAQUIGRÁFICAS} site given a main word or phrase and the dates.
}
\keyword{internal}
