% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-speech-table.R
\name{speech_data}
\alias{speech_data}
\title{Obtains the text of the speech}
\usage{
speech_data(
  keyword,
  reference_date,
  qtd_days,
  uf = "",
  orador = "",
  partido = ""
)
}
\arguments{
\item{keyword}{the keyword used to search speeches.}

\item{reference_date}{end date of search.}

\item{qtd_days}{quantity of days before the \code{reference_date}.}

\item{uf}{state acronym.}

\item{orador}{speaker's name.}

\item{partido}{political party of speaker.}
}
\value{
the speech data with all informational columns and the speech.
}
\description{
Extract the text of the speech given an URL.
}
\examples{
speech_data(keyword = '"meio ambiente"', reference_date = "2021-12-20", qtd_days = 3)
}
