% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-speech-table.R
\name{speech_data}
\alias{speech_data}
\title{Obtains the text of the speech}
\usage{
speech_data(keyword, reference_date, qtd_days)
}
\arguments{
\item{keyword}{argument of the \code{download_page} function.}

\item{reference_date}{end date of search, the default is the last day of 2021.}

\item{qtd_days}{quantity of days before the \code{reference_date}, the default is 5 days.}
}
\value{
vector the speech data with all informational columns and the speech.
}
\description{
Extract the text of the speech given an URL.
}
\examples{
speech_data(keyword = '"meio ambiente"', reference_date = "2021-12-20", qtd_days = 5)
}
